+++ messaging-ui-tests/src/main/java/com/nabla/sdk/uitests/scene/MainActivity.kt
@@ -1,26 +1,11 @@
 package com.nabla.sdk.uitests.scene
 
-import android.content.Intent
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.lifecycleScope
-import com.nabla.sdk.messaging.ui.helper.ConversationListViewModelFactory
-import com.nabla.sdk.messaging.ui.scene.conversations.ConversationListViewModel
-import com.nabla.sdk.messaging.ui.scene.conversations.bindViewModel
+import androidx.fragment.app.Fragment
 import com.nabla.sdk.uitests.databinding.ActivityMainBinding
-import com.nabla.sdk.uitests.scene.ConversationActivity.Companion.CONVERSATION_ID_EXTRA
-import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
-
-    private val viewModel: ConversationListViewModel by viewModels {
-        ConversationListViewModelFactory(
-            owner = this,
-            messagingClient = nablaMessagingClient
-        )
-    }
-
     private lateinit var binding: ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -28,21 +13,19 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        lifecycleScope.launch {
-            binding.createConversation.setOnClickListener {
-                launch {
-                    nablaMessagingClient.createConversation()
-                }
-            }
-            binding.conversationListView.bindViewModel(
-                viewModel,
-                onConversationClicked = { id ->
-                    startActivity(
-                        Intent(applicationContext, ConversationActivity::class.java)
-                            .apply { putExtra(CONVERSATION_ID_EXTRA, id.value) }
-                    )
-                }
-            )
+        if (savedInstanceState == null) {
+            supportFragmentManager
+                .beginTransaction()
+                .replace(binding.fragmentContainer.id, ConversationsFragment())
+                .commit()
         }
+    }
+
+    fun pushFragment(fragment: Fragment) {
+        supportFragmentManager
+            .beginTransaction()
+            .replace(binding.fragmentContainer.id, fragment)
+            .addToBackStack(null)
+            .commit()
     }
 }