+++ messaging-ui-tests/src/main/java/com/nabla/sdk/uitests/scene/StubbedConversionFragment.kt
@@ -1,0 +1,7 @@
+package com.nabla.sdk.uitests.scene
+
+import com.nabla.sdk.messaging.ui.scene.messages.ConversationFragment
+
+class StubbedConversionFragment : ConversationFragment() {
+    override val messagingClient = nablaMessagingClient
+}